Cleaned up include dependencies on Linux through use of pkg-config. Integrated changes from Mark Wieder and Matt Peterson to get Linux running under 64-bit. Integrated Makefile changes from Matt Peterson (added 'clean' target and added missing dependencies).

+++ engine/src/sysdefs.h
@@ -708,9 +708,9 @@ typedef unsigned long Window;
 typedef unsigned long Pixmap;
 typedef unsigned long Drawable;
 #else
-typedef unsigned int Window;
-typedef unsigned int Pixmap;
-typedef unsigned int Drawable;
+typedef unsigned long int Window;
+typedef unsigned long int Pixmap;
+typedef unsigned long int Drawable;
 #endif
 
 #endif
